Moved gui files to appropriate subdirectories
