+++ docs/source/changelog.md
@@ -276,7 +276,7 @@ Below is a conceptual before/after to illustrate the change.
 
 ```python
 # example (conceptual)
-from algomancy_gui.appconfiguration import AppConfiguration
+from algomancy_gui.configuration.appconfiguration import AppConfiguration
 
 
 class ExampleDataPage:
@@ -298,7 +298,7 @@ config = AppConfiguration(
 
 ```python
 # example (conceptual)
-from algomancy_gui.appconfiguration import AppConfiguration
+from algomancy_gui.configuration.appconfiguration import AppConfiguration
 
 
 class ExampleDataPage:
@@ -634,17 +634,17 @@ Your main method must be migrated to use the new class. An example is shown belo
 # main method: preferred version
 
 from algomancy_gui.gui_launcher import GuiLauncher
-from algomancy_gui.appconfiguration import AppConfiguration
+from algomancy_gui.configuration.appconfiguration import AppConfiguration
 
 
 def main():
-    app_cfg = AppConfiguration(
-        data_path="data",
-        has_persistent_state=True,
-        #       ...
-    )
-    app = GuiLauncher.build(app_cfg)
-    GuiLauncher.run(app=app, host=app_cfg.host, port=app_cfg.port)
+  app_cfg = AppConfiguration(
+    data_path="data",
+    has_persistent_state=True,
+    #       ...
+  )
+  app = GuiLauncher.build(app_cfg)
+  GuiLauncher.run(app=app, host=app_cfg.host, port=app_cfg.port)
 ```
 For migration, the `AppConfiguration.from_dict(...)` method can be used to create an `AppConfiguration` object from a dictionary. Note that this is not advised, as this will not allow for IDE support.
 
@@ -652,18 +652,18 @@ For migration, the `AppConfiguration.from_dict(...)` method can be used to creat
 # main method: migration alternative
 
 from algomancy_gui.gui_launcher import GuiLauncher
-from algomancy_gui.appconfiguration import AppConfiguration
+from algomancy_gui.configuration.appconfiguration import AppConfiguration
 
 
 def main():
-    configuration = {
-        "data_path": "data",
-        "has_persistent_state": True,
-        #       ...   
-    }
-    app_cfg = AppConfiguration.from_dict(configuration)
-    app = GuiLauncher.build(app_cfg)
-    GuiLauncher.run(app=app, host=app_cfg.host, port=app_cfg.port)
+  configuration = {
+    "data_path": "data",
+    "has_persistent_state": True,
+    #       ...   
+  }
+  app_cfg = AppConfiguration.from_dict(configuration)
+  app = GuiLauncher.build(app_cfg)
+  GuiLauncher.run(app=app, host=app_cfg.host, port=app_cfg.port)
 ```
 ### Autocreate
 Added automatic creation of scenarios. This will cause any creation of a `DataSource` (or derived) to spawn a `Scenario` with the same name (suffixed with `[auto]`). The algorithm template must be specified in the configuration dictionary.
