+++ packages/algomancy-gui/src/algomancy_gui/configuration/stylingconfigurator.py
