+++ packages/algomancy-gui/src/algomancy_gui/data_page/data.py
@@ -9,9 +9,9 @@ from ..componentids import (
     DATA_PAGE,
 )
 from ..data_page.datamanagementtopbar import top_bar
-from algomancy_gui.managergetters import get_scenario_manager
-from ..layouthelpers import create_wrapped_content_div
-from ..contentregistry import ContentRegistry
+from algomancy_gui.managers.managergetters import get_scenario_manager
+from algomancy_gui.loaders.contentwrapper import create_wrapped_content_div
+from algomancy_gui.managers.contentregistry import ContentRegistry
 
 from ..data_page import dialogcallbacks  # noqa
