+++ packages/algomancy-gui/src/algomancy_gui/data_page/datamanagementdownloadmodal.py
@@ -22,7 +22,7 @@ import tempfile
 import threading
 
 from algomancy_scenario import ScenarioManager
-from algomancy_gui.managergetters import get_scenario_manager
+from algomancy_gui.managers.managergetters import get_scenario_manager
 
 """
 Modal component for downloading data files into the application.
