+++ packages/algomancy-gui/src/algomancy_gui/gui_launcher.py
@@ -8,12 +8,12 @@ from dash import get_app, Dash, html, dcc
 from dash_bootstrap_components.themes import BOOTSTRAP
 
 from .layout import LayoutCreator
-from .contentregistry import ContentRegistry
-from .settingsmanager import SettingsManager
-from .sessionmanager import SessionManager
+from algomancy_gui.managers.contentregistry import ContentRegistry
+from algomancy_gui.managers.settingsmanager import SettingsManager
+from algomancy_gui.managers.sessionmanager import SessionManager
 from .componentids import ACTIVE_SESSION
-from .appconfiguration import AppConfiguration
-from .librarymanager import LibraryManager as lm
+from algomancy_gui.configuration.appconfiguration import AppConfiguration
+from algomancy_gui.managers.librarymanager import LibraryManager as lm
 from algomancy_scenario import ScenarioManager
 from algomancy_utils.logger import MessageStatus
