+++ packages/algomancy-gui/src/algomancy_gui/managers/contentregistry.py
