+++ packages/algomancy-gui/src/algomancy_gui/managers/librarymanager.py
