+++ packages/algomancy-gui/src/algomancy_gui/managers/managergetters.py
