+++ packages/algomancy-gui/src/algomancy_gui/managers/settingsmanager.py
