+++ packages/algomancy-gui/src/algomancy_gui/scenario_page/new_scenario_creator.py
@@ -9,12 +9,12 @@ from algomancy_gui.componentids import (
     SCENARIO_CREATE_STATUS,
     SCENARIO_CREATOR_MODAL,
 )
-from algomancy_gui.managergetters import get_scenario_manager
+from algomancy_gui.managers.managergetters import get_scenario_manager
 from algomancy_gui.scenario_page.new_scenario_parameters_window import (
     create_algo_parameters_window,
 )
 from algomancy_scenario import ScenarioManager
-from algomancy_gui.stylingconfigurator import StylingConfigurator
+from algomancy_gui.configuration.stylingconfigurator import StylingConfigurator
 
 
 def new_scenario_creator(session_id: str):
