+++ packages/algomancy-gui/src/algomancy_gui/scenario_page/scenarios.py
@@ -47,10 +47,10 @@ from .new_scenario_parameters_window import (
     create_algo_parameters_entry_card_body,
 )
 from .scenario_cards import scenario_cards
-from ..contentregistry import ContentRegistry
-from algomancy_gui.managergetters import get_scenario_manager, get_manager
+from algomancy_gui.managers.contentregistry import ContentRegistry
+from algomancy_gui.managers.managergetters import get_scenario_manager, get_manager
 
-from ..layouthelpers import create_wrapped_content_div
+from algomancy_gui.loaders.contentwrapper import create_wrapped_content_div
 from .delete_confirmation import (
     delete_confirmation_modal,
 )
@@ -60,7 +60,7 @@ import dash_bootstrap_components as dbc
 
 from .scenario_cards import hidden_card
 from algomancy_scenario import ScenarioManager
-from ..settingsmanager import SettingsManager
+from algomancy_gui.managers.settingsmanager import SettingsManager
 
 
 def scenario_page():
